refactor(eo): use `[eoMediaImage]` selector
